feat: replace OpticModifier with OpticUpdater for enhanced optical system updates
